# openapi/control.yaml
openapi: 3.1.0
info:
  title: Eugene Plexus — control
  version: 0.3.0-draft
  summary: The install's trust root, node registry, topology and UI host. Exactly one active.
  description: |
    The **control root**. Exactly one is active per install, plus any
    number of warm standbys. It holds what is inherently singular —
    the trust root, the node registry, install-wide topology, the
    replicated log and the UI — and it **spawns nothing**.

    Supervision is the [agent](agent.yaml)'s job, and there is one agent
    per host. On whichever host holds the control root, `control` is
    itself just another local component in that agent's topology. That
    is deliberate: if the control root supervised processes it would need
    a second copy of the supervision machinery, and "components share
    schemas, not code" means a real second copy rather than an import.
    One supervisor implementation, running everywhere; one control root,
    supervised like anything else.

    Design and every decision behind this document:
    `docs/design/m5-multi-host-and-trust.md`. Implemented in the
    `control` repo; the agent's half of the exchange — enrollment, an
    address other hosts reach it at — landed at M7
    (`docs/design/m7-second-host-readiness.md`), and per-node token keys
    with the trust bundle on 2026-09-25
    (`docs/design/per-node-token-keys.md`).

    ## What it is authoritative for

    | | control root | agent |
    |---|---|---|
...

...

    ## `down` versus `out`
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/people:
    get:
      operationId: listPeople
      summary: The people who may sign in to apps
      description: Operator only. Never a password or its verifier.
      responses:
        '200':
          description: The people.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PersonList'}
    post:
      operationId: createPerson
      summary: Add a person
      description: Operator only; active, unlocked root. A name is unique, compared case-folded, and `operator` is the owner's.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PersonCreateRequest'}
      responses:
        '201':
          description: The person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Person'}
        '409':
          $ref: 'components/common.yaml#/components/responses/Problem'
        '422':
          $ref: 'components/common.yaml#/components/responses/Problem'
...
  /v1/oidc/clients:
    get:
      operationId: listOidcClients
      summary: The apps that sign in with Eugene
      description: Operator only. Never a secret.
      responses:
        '200':
          description: The apps.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OidcClientList'}
    post:
      operationId: createOidcClient
      summary: Register an app that signs in with Eugene
      description: |
        Operator only, or a member node acting for the operator on its
        own app (`SubjectToken`, with `owner` `app:<id>@<node>`); active,
        unlocked root. Redirect URIs are matched exactly (RFC 9700 §4.1);
        plain `http` is allowed, because a home network has no
        certificates. The secret is in this answer and nowhere else.
      parameters:
        - $ref: '#/components/parameters/SubjectToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OidcClientCreateRequest'}
      responses:
        '201':
          description: The app and its one-time secret.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OidcClientCreated'}
        '403':
          $ref: 'components/common.yaml#/components/responses/Problem'
        '422':
          $ref: 'components/common.yaml#/components/responses/Problem'
...
  /v1/auth/initialize:
    post:
      tags: [auth]
      operationId: authInitialize
      summary: Set the operator passphrase on a fresh install.
      security: []
      description: |
        First-run only, and refused once `initialized` is true. Derives
        the master key via Argon2id and generates this root's token
        key and its epoch-1 identity.

        The passphrase is unrecoverable by design: losing it means losing
        every sealed secret in the install, including the recovery
        recipient that would otherwise reach a dead node's credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthInitializeRequest'
      responses:
        '204':
          description: Initialized.
        '409':
          description: Already initialized.
          content:
            application/problem+json:
              schema:
                $ref: 'components/common.yaml#/components/schemas/Problem'
...
  /oidc/node-helpers/execute:
    post:
      operationId: executeNodeFileOperation
      tags: [node-helpers]
      summary: executeNodeFileOperation
      description: An enrolled Workbench client secret and its signed-in person refresh credential. Current grants are checked for every call.
      security: [{helperClient: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelperCall"
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        '401':
          description: Authentication, current permission, state, or availability refused the request.
        '403':
          description: Authentication, current permission, state, or availability refused the request.
        '409':
          description: Authentication, current permission, state, or availability refused the request.
        '503':
          description: Authentication, current permission, state, or availability refused the request.
...
components:
  securitySchemes:
    helperClient:
      type: http
      scheme: basic
      description: Workbench client_secret_basic, plus that client's person refresh token in the body.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        An operator session from `POST /v1/auth/login` or
        `POST /v1/auth/token`, or a service token addressed to `control`
        and signed by a key in the trust bundle (`common.yaml`,
        `TrustBundle`). Mutations take a session; the reads that say
        so also take an `agent` or `gateway` service token.
  parameters:
    SubjectToken:
      name: X-Eugene-Plexus-Subject-Token
      in: header
      required: false
      schema: {type: string}
      description: |
        **A member node acting for an operator who is acting on it**
        (2026-10-01). The console reaches another machine with a
        five-minute token addressed to that machine alone (`POST
        /v1/auth/token`), and installing an app there needs its key and
        its sign-in registration from this root -- which that token, not
        being addressed here, cannot ask for. So the machine sends its
        own `agent` service token as the bearer (the actor) and the
        operator's token, addressed to that same machine, here (the
        subject; RFC 8693's roles).

        Taken only by the operations that reference it, and only for
        things named for the acting node: a client key named
        `app:<id>@<node>`, a sign-in registration whose `owner` is. Any
        other name is a 403. A subject addressed to another machine, an
        actor that is not an agent speaking for itself, or a signed-out
        session is a 401. A node's own token without a subject opens
        none of these.
  schemas:
    HelperGrant:
      $ref: "components/node-helpers.yaml#/HelperGrant"
...
    HelperCall:
      $ref: "components/node-helpers.yaml#/HelperCall"
...
    Person:
      type: object
      description: |
        Someone the operator lets sign in to apps (C2). Not an operator:
        a person's sign-in opens their apps and nothing in the hub.
      required: [id, name, disabled, createdAt, passwordChangedAt]
      properties:
        helperGrants:
          type: array
          maxItems: 256
          description: Explicit access to folders on enrolled nodes; absent means no access.
          items: {$ref: "#/components/schemas/HelperGrant"}
        id: {type: string, description: 'Stable; the `sub` in their ID tokens.'}
        name: {type: string, description: 'What they sign in with. Unique, compared case-folded.'}
        displayName: {type: string}
        email:
          type: string
          description: |
            Optional, and unique compared case-folded. Given to an app that
            asks for the `email` scope, with `email_verified: false`: Eugene
            sends no mail and has proved nothing about it. Some apps (Open
            WebUI) identify people by it (C4).
        apps:
          type: [array, 'null']
          items: {type: string}
          description: The `clientId`s they may sign in to. Null is every app on the install.
        disabled: {type: boolean}
        createdAt: {type: string, format: date-time}
        passwordChangedAt: {type: string, format: date-time}
    PersonList:
      type: object
      required: [people]
      properties:
        people:
          type: array
          items: {$ref: '#/components/schemas/Person'}
        operatorName:
          type: string
          description: The name the owner signs in with, once the install has people.
    PersonCreateRequest:
      type: object
      additionalProperties: false
      required: [name, password]
      properties:
        helperGrants:
          type: array
          maxItems: 256
          description: Explicit access to folders on enrolled nodes; absent means no access.
          items: {$ref: "#/components/schemas/HelperGrant"}
        name: {type: string, minLength: 1, maxLength: 64, pattern: "^[^\\s@][^@]*[^\\s@]$|^[^\\s@]$"}
        displayName: {type: string, maxLength: 120}
        email: {type: string, maxLength: 254, pattern: "^[^@\\s]+@[^@\\s]+$"}
        password: {type: string, minLength: 12, maxLength: 1024}
        apps:
          type: [array, 'null']
          items: {type: string}
...
    OidcClient:
      type: object
      required: [clientId, name, redirectUris, createdAt]
      properties:
        clientId: {type: string}
        name: {type: string}
        redirectUris:
          type: array
          items: {type: string}
        owner:
          type: string
          description: '`app:<id>@<node>` for an app the registry installed; absent for one the operator added.'
        createdAt: {type: string, format: date-time}
    OidcClientList:
      type: object
      required: [clients]
      properties:
        clients:
          type: array
          items: {$ref: '#/components/schemas/OidcClient'}
        issuer:
          type: string
          description: '`oidcIssuer` when the operator set one; otherwise absent, and an app uses the address it reaches Eugene at, with `/oidc`.'
    OidcClientCreateRequest:
      type: object
      additionalProperties: false
      required: [name, redirectUris]
      properties:
        name: {type: string, minLength: 1, maxLength: 80}
        redirectUris:
          type: array
          minItems: 1
          maxItems: 20
          items: {type: string, minLength: 1, maxLength: 2048}
        owner: {type: string, maxLength: 120}
    OidcClientCreated:
      type: object
      required: [client, clientSecret]
      properties:
        client: {$ref: '#/components/schemas/OidcClient'}
        clientSecret:
          type: string
          description: In this answer only.
...
    AuthInitializeRequest:
      type: object
      required:
        - passphrase
      properties:
        passphrase:
          type: string
          minLength: 1
          format: password
          description: |
            Hashed with Argon2id and used to derive the master key that
            at-rest secrets are sealed under. Losing it means losing
            every stored credential — including the recovery recipient
            that would otherwise reach a dead node's secrets. There is no
            recovery path, by design.